Make guide design be deterministic
Previously, when making consensus sequences, ties in base
counts at a site would be broken by the max(..) function.
This would choose the base that appeared first in the
dictionary it was given, but the order of that dictionary was
arbitrary and could vary across runs. Thus, the consensus
sequence could vary across runs when there were ties, which
means that the guide sequences (and counts within a window)
could vary across runs.

This changes the behavior when there are ties by giving
the max(..) function a list of tuples sorted by base. Thus,
it breaks ties in base count at a site by choosing the
base that is first alphabetically. This is still arbitrary,
but is now deterministic. It means that the guide design
is now deterministic.