Create output file that details guide sequences by window
Previously, the design tool simply output summary information
of the guide sequences -- it selected the "best" window (based
on number of guides sequences and a score), broke ties arbitrarily,
and output the guides for this window. This change now requires
an output TSV file as an argument, and it writes to this file
the guide sequences that are designed for each window.

That way, a user can have all the information about the guides
across the windows and decide for themselves which window to
use. It can optionally sort (with `--sort`) the file so that
the "best" choices of window are on top.

The tool now writes to stdout a brief analysis that summarizes
the design across the windows.