Improve handling of sequences with gaps
Previously, guides would not be designed to target a region (e.g., a 28-nt
stretch) if any sequence in the alignment at that region contained a gap. Now,
it will try to design guides in such regions, but will not construct guides
based on sequences that contain gaps and will not attempt to cover such
sequences.