Impose additional threshold on when to ignore sites with missing data
The median fraction of sequences with missing data over the genome can easily
be 0 (i.e., if there is little missing data). Previously, this would then
preclude guide design in a region in which _any_ sequence has missing data
(because the fraction of sequences with missing data would exceed 0). This is
arguably too harsh. This commit adds a minimum b such that if the median (times
a constant factor) is < b, then b is used instead.