Cluster potential guide sequences before taking consensus
Previously, we would approximate an optimal guide sequence to cover a set S of
sequences by taking the consensus over S of the region that the guide would
target. This commit changes how the approximation is done. Now, it produces a
rough clustering of the sequences in S (just the guide region) using LSH, and
takes the consensus over only the largest cluster. This can lead to either more
or fewer guides depending on the actual sequences in S, but the hope is that,
in practice, this will lead to fewer guides.