Mask current alignment from specificity checks when designing for differential identification
Previously, when designing guides to target sequences in alignment A_i for use
in differential identification, this would do a near neighbor lookup for each
possible guide g against all sequences (across all alignments) that contain a
portion whose sequence is near (similar to) g. It does this to check the
specificity of g. Almost all reports for the query of g is sequence in A_i,
which is irrelevant for specificity. Thus, this would filter out all reports of
sequence from A_i after doing the query. However, this is extremely slow: the
query() method of NearNeighborLookup would calculate the distance between g and
each possible matching sequence in A_i to see if it should be reported.
Calculating these distances across so many possible matchings is slow.

This commit adds methods to NearNeighborLookup for masking points, and applies
them to mask all sequence from A_i when designing guides for A_i. This way,
sequence from A_i does not show up in the hash tables when querying.  So there
are many fewer possible points to report (and against which to calculate
distance), making queries much faster.