Generalize universe to encompass multiple groups
When finding desired coverage of elements in a universe, this generalizes the
universe so that it can consist of separate groups of elements. (Another way to
think about this is that it searches over multiple universes.) Each group in
the universe can have its own desired partial cover to achieve. This commit
updates the functions that solve a set cover instance so support an instance
that is generalized in this way.

As one example, we could imagine that each group in the universe consists of
sequences from a particular year. Then, we could use this to achieve more
partial coverage across sequences from recent years and less coverage across
sequences from more distant years.