Construct guides while accounting for the number of sequences needed to achieve a partial cover
While constructing guides over a universe of sequences that can be divided into
groups, it is important to account for the number of sequences needed to
achieve the desired coverage in each group. Imagine we have groups of sequences
from 2 years, 2010 and 2018, and we would like to achieve 10% coverage across
sequences from 2010 but 100% from sequences across 2018. If there are many more
sequences from 2010 than there are from 2018, simply taking the consensus
across all of these sequences would weight more heavily toward the 2010
sequences; therefore, it may not cover the 2018 sequences well and may provide
sub-optimal coverage.

This modifies the function that constructs a guide (along with its wrapper to
memoize output) to, optionally, incorporate the number of sequences from each
universe group that are left to cover in order to achieve the desired partial
cover. It clusters the sequences needed to be covered (across all groups,
regardless of how much desired partial cover each has), and then constructs a
consensus sequence across only the cluster A, where A is the cluster that
contains the most number of sequences that are needed to be covered while
accounting for how many sequences are left to cover in each group.