Calculate score in window based on fraction of needed sequences
Previously, a score for a guide set (i.e., in a window) was calculated based on
the fraction of all sequences that the guides covered. This adjusts the score
calculation to be based on the fraction of all sequences that *need* to be
covered (according to the desired partial cover) that are covered by the
guides.

This helps account for partial coverage of sequences that varies by year. That
is, if sequences from a distant year are not covered well by the guides (and
these sequences do not *need* to be covered well), this will not penalize the
score by much.  Instead, the score will depend more on coverage of sequences
from a recent year (that may *need* to be covered well).