Consider >1 cluster from which to construct guides
Previously, alignment.Alignment.construct_guide(..) only considered
a single cluster of sequences (the "best" one according to a heuristic,
or the largest) and constructed a guide from the consensus of these
sequences. But this guide could later be deemed not suitable (if
guide_is_suitable_fn(guide) is False) and ignored in
guide_search.GuideSearcher._find_optimal_guide_in_window(..).

This commit changes alignment.Alignment.construct_guide(..) to
directly take the guide_is_suitable_fn function. It ranks the
clusters by a heuristic, and then designs a guide from the
consensus of each, picking the one from the "best" cluster
such that the guide it yields is suitable.

This should help in designing guides for differential identification
between related sequence, when it may be useful to construct
guides from less-than-ideal clusters in order to find ones that are
sufficiently specific.