Add option to GuideSearcher to require certain guides in its output
This adds a required_guides argument to the initializer of
GuideSearcher. The argument provides a set of guide sequences
(and their positions in the alignment) that must be in the
output for the windows that contain them. This might be useful,
for example, if certain guides have been shown experimentally
to perform well. _find_guides_that_cover_in_window(..) initializes
its output with these guide sequences, and updates the universe
accordingly, so that the sequences they cover are already
accounted for.

This also adds test cases for this option.