Add option to GuideSearcher to blacklist ranges
This adds a blacklisted_ranges argument to the initializer of
GuideSearcher. The argument provides a set of blacklisted ranges
in the alignment, such that no guides overlapping these ranges
should be constructed. This might be useful, for example, if
certain ranges in an alignment contain sequence that is hard
to detect (e.g., due to structure). _find_optimal_guide_in_window(..)
avoids constructing any guide that overlaps these ranges.

This also adds test cases for this option.