Add --blacklisted-kmers argument
This updates design_guides.py executable to accept a
--blacklisted-kmers argument. The argument provides a
path to a FASTA file that gives k-mers such that no
guides should be constructed that contain these k-mers.

This updates seq_io.py to parse that file.

This commit implements k-mer blacklisting through the
guide_is_suitable_fn argument to GuideSearcher. It now
passes a guide_is_suitable_fn to GuideSearcher when
designing guides independently, where the function
decides a guide is suitable iff it does not contain a
blacklisted k-mer. It also updates the guide_is_suitable_fn
when designing guides for ID to check whether the guide
contains a blacklisted k-mer, and return False if it
does.