Tolerate G-U base pairing when determining whether guide binds to target
This adds an option to the guide module that specifies
whether to allow G-U base pairing when determining
whether a guide binds to a target. It is enabled by
default. Because many unit tests were written without
tolerating G-U base pairing (and would break with this),
mosts tests are now run with this option disabled. This
commit also adds several new unit tests with this
option enabled.

This commit also updates the design_guides.py executable
with an argument '--do-not-allow-gu-pairing' that, when
set, disables the above option.