Add HammingWithGUPairsDistanceFamily to allow LSH tolerating G-U base pairs
This adds an LSH hash family, HammingWithGUPairsDistanceFamily, that
works with a pseudo-Hamming distance, which tolerates G-U base pairing.

The comment in the HammingWithGUPairsDistanceFamily implementation
provides information about how this works.

This family can be used to support near-neighbor queries whose output
includes sequences similar to the queried sequence, accounting
G-U base pairs, provided that the distance function used with those
queries accurately measures distance between a guide and target
sequence (unlike the distance function implemented in this commit).
It should be used with caution when clustering sequences, as explained
in the comment for HammingWithGUPairsDistanceFamily.

This commit also includes new unit tests for near-neighbor
queries that account for G-U base pairs. These tests use the
guide.seq_mismatches_with_gu_pairs(..) function as the distance
function.