Perform queries in AlignmentQuerier against gapless sequences
This adds an optional argument to alignment.Alignment.make_list_of_seqs()
to remove gaps ('-') from the sequence strings it returns.

This also updates alignment.AlignmentQuerier.setup() to use this
option when constructing its self.nnr data structure. As a result,
the indexed subsequences (in particular, all l-mers of the sequences
where l is the guide length) will be gapless. When querying for
potential hits against the sequences, this will allow for a guide to
be detected as hitting a subsequence even if that subsequence has
gaps in the alignment (e.g., the guide 'GATACA' can hit the sequence
'GA--TACA' because its gaps will have been removed prior to being
stored in self.nnr).

Note also that, without this change, lsh.NearNeighborLookup.query()
can run into a problem if the sequence being queried against
contains gaps. This is because if its self.dist_fn is
guide.seq_mismatches() or guide.seq_mismatches_with_gu_pairs(),
it will crash because those functions have a precondition that
the sequences provided to them do not contain gaps.