Fix bug caused by editing seqs_to_consider in construct_guide()
The dict `seqs_to_consider` is passed by reference to Alignment.construct_guide(),
so it points to the same dict as `universe` in
GuideSearcher._find_guides_that_cover_in_window() (this calls
_find_optimal_guide_in_window() with universe, which calls
_construct_guide_memoized() with that same dict, which calls
Alignment.construct_guide() with that same dict).

Alignment.construct_guide() directly modified the argument `seqs_to_consider`
that was passed to it. As a result, this also modified `universe` in
GuideSearcher._find_guides_that_cover_in_window(). This is incorrect and
has dangerous effects. For example, because Alignment.construct_guide() removes
sequences with a gap from `seqs_to_consider`, this effectively removes those
same sequences from `universe` in GuideSearcher._find_guides_that_cover_in_window().
In some cases, this could mean that sequences with a gap never need to be
covered (and thus a cover is returned that is presumed to achieve the desired
coverage, but does not).

This commit fixes the bug by making a copy of `seqs_to_consider` in
Alignment.construct_guide() before modifying it.

This commit also includes a new unit test that would have failed with this bug,
but now passes.