Add module for curating and aligning sequences
To perform alignment, this module requires a path to mafft,
the program it calls. The alignment is executed by the align()
function.

This performs curation of sequences, through the curate_against_ref()
function, by aligning each sequence against a specified reference
sequence and checking identity.

This module also includes two classes for memoizing alignments
and statistics on alignments by storing them to files. Because
many uses of this module will be on the same set of sequences as
previous sessions, and alignment can be an expensive operation,
this can potentially save considerable resources across different
sessions.