Make GuideSearcher more flexible with window size
Previously, guide_search.GuideSearcher was initialized with a fixed
window size and all its methods used this window size. This commit
updates GuideSearcher so that methods accept windows with any
specified start/end coordinates -- i.e., so that the size of the
window is flexible. To keep compatibility with runs of a fixed
window size (sliding across the alignment), this changes the
high-level methods to accept an argument that gives the window
size, and these methods then pass along start/end coordinates
according to the window size.