Set G-U pairing on each function call, rather than globally
Previously, the option of whether to allow G-U pairing was set
globally in the guide module. This was simpler, but would make
it difficult to run GuideSearcher in parallel with different
settings. It would also complicate using PrimerSearcher (with
G-U pairing off) simultaneously with GuideSearcher (with G-U
pairing on).

This commit sets the option when initializing a GuideSearcher
object (and similar objects), and passes the setting along to
lower function calls. This is messier, but has the advantages
listed above.