Only design guides across sequences that are bound by primers
This change only affects design by target_search: When constructing
guides within a window bound by a pair of primer sets, this finds
the sequences that would be amplified by the primers. Then, it
only designs guides across those sequences by initializing the
universe to contain only these.

For example, suppose that cover_frac for primer_search were 0.5
and cover_frac for guide_search were 0.5. It is possible in
theory that primers are designed to amplify 50% of the sequences,
but the guides between those amplify the *other* 50% of sequences.
This change alleviates this problem. Now, 50% *of the sequences
amplified by the primers* are covered by guides, which might
potentially be just 25% of all sequences. It is reasonable,
with this change, to set cover_frac for guide_search to be 1.0
(or close to it) with cover_frac for primer_search being lower;
this way, nearly all or all of the sequences amplified by the
primers will be covered by guides.