Add option to find targets while collapsing overlapping ones
Often, many of the best N targets will be highly similar in position:
the primers on their ends will overlap. This considers two targets
to overlap if the primers on the ends both overlap.

This commit adds an argument to find_targets() that will prevent
the results from containing overlapping targets. In particular, if
two targets overlap, only the one with the lowest cost is kept
in the output. This determines overlapping ones, and chooses whether
to replace an existing target that overlaps a new one, during the
search so that N targets are still output.