Add argument to limit target length in the search
Targets that are especially long may be searched (i.e., guides
looked for within them) early in the search, when target_heap
is not yet full. It can be slow to find guides within these
large windows. Specifying a maximum target length allows more
potential targets to be pruned early on, speeding up the
search.