Change data structures for alignment queries to hold less repetitive information
Previously, the NearNeighborLookup object (nnr) in AlignmentQuerier stored
tuples (subsequence s, alignment i, sequence j), key'd on s. A particular
subsequence could appear could appear in many (if not most or all) sequences
in an alignment, and thus there could be many tuples in nnr representing
a unique s. This hurts memory usage because a single s, particularly if
present in many sequences (e.g., from a conserved region), appears many
times in nnr's hashtables. It also hurts runtime because a query for a
neighbor of s (i.e., a guide sequence similar to s) will yield many
possible neighbors p_key in NearNeighborLookup.query(), and that function
will have to evaluate each of them (by calling dist_fn(q, p_key), which
is slow).

However, we do not need to store that complete tuple in nnr for
AlignmentQuerier to query potential guide sequences.

This changes AlignmentQuerier to only store (subsequence s, alignment i)
in nnr. It also adds a data structure to AlignmentQuerier that stores
all of the sequences j for each pair (s, i) that contain s and are in
alignment i. Thus, when nnr.query(q) returns near-neighbors of q (the
tuples (s, i)), AlignmentQuerier can find the particular sequences in
i that contain s.