Increase default k for AlignmentQuerier to k=22
An significant upside to increasing k is that it will improve
runtime because there can be far (exponentially) fewer false
positives that must be evaluated in NearNeighborLookup.query().
The main downside is that it raises memory usage because
NearNeighborLookup will use more hashtables for a particular choice
of reporting probability.

The parent commit should permit less memory usage. As a result,
this increases k to achieve an improved runtime.