Add missing guide suitability check at end of construct_guide()
This adds a missing suitability check for guide sequences at
the end of construct_guide(), in the edge case where a guide
does not bind to any sequences.

This also removes a (double-)check on the guide suitability
in guide_search. This second check was unnecessary; since
running guide_is_suitable_fn can be slow, this commit removes
the second check.