Implement required flanking seqs argument in Alignment functions
Two functions in Alignment determine which sequences a guide
binds to: construct_guide() and sequences_bound_by_guide(). Both
should account for the required flanking sequences. This commit
adds required_flanking_seqs as an argument to these functions,
and only decides that guides can bind to sequences in an alignment
that have the required sequence flanking the guide.

In construct_guide(), it additionally accounts for required_flanking_seqs
in the design: it only designs against sequences in the alignment
that have the required flanking sequence around the site of the
guide, effectively treating the others as a lost cause.