Change MinHash hash function to return signature
Previously, the MinHash hash function returned just a single value
(the minimum value over the k-mers of a sequence). This commit
changes it to return the N smallest, for a given N.

This also adds a function to estimate the Jaccard distance between
two sets of k-mers (sequences), given their MinHash signatures.