Cluster genomes prior to alignment with auto- subcommands
This modifies prepare_alignment.prepare_for() to cluster the input
genomes, and separately align the genomes in each cluster. The
clustering is only performed when using the
{auto-from-file,auto-from-args} subcommands (i.e., not when given an
alignment as a FASTA file).

This also modifies the design_guides.py executable to use separate
alignments for each taxon. It keeps track of the different alignments
for a taxon (i.e., the taxid for each alignment) to use with
differential identification. As a result, this also requires modifying
AlignmentQuerier to perform specificity checks against not only a single
alignment, but also a collection of alignments (i.e., all the alignments
in the same taxon as the one for which guides are being designed for).