Allow curation against multiple reference sequences
For some taxa (e.g., influenza A virus segment 4), a single reference
sequence used for curation cannot capture the diversity that should be
included.  This commit allows curating against multiple reference
sequences, such that a sequence s is included in the design if s
satisfies the criteria (is similar enough to) any of the reference
sequences.

The reference sequences are provided in a comma-separated list.