Allow variable cover_frac, by grouped sequences, in primer_search
This commit modifies primer_search.PrimerSearcher to accept
a collection of sequence groups (seq_groups) and also a
particular desired cover fraction for each, as done in
guide_search.GuideSearcher.

This commit also modifies design.py to implement the
cover-by-year-decay feature for primer searching. As a result,
when searching for complete targets, cover-by-year-decay is
now used for both primers and guides.