Allow AlignmentStatMemoizer to be used in parallel
Previously, AlignmentStatMemoizer stored stats on alignments in a single
file. This could prevent its use by different processes at the same time
(assuming they all memoize to the same file).

This commit changes AlignmentStatMemoizer to be more like
AlignmentMemoizer, storing stats for each alignment in a separate file.

The primary downside, and the reason this was not done initially, is
that it may result in a very large number of files being created because
a save() is performed for each curation; in effect, there will be one
file for each sequence (since stats on each sequence-reference alignment
will be saved).