Improve logging during input preparation
When sampling sequences with replacement, the number of sequences
(and number of unique accessions) that are output at each step can
be confusing. This is in part because curation is only done on
unique accessions, and then repeated accessions (i.e., that were
sampled more than once) are added in (given multiplicity) after
curation.

This commit aims to improve some of the log messages to clarify
the state.