Add design_naively.py executable
The output of this script can be used as a baseline for design.py.

This designs guides naively. For each window (sliding over the
alignment), it constructs guides two ways: a single guide
based on the consensus and a single guide based on the mode
(most common sequence). It outputs a TSV file giving the best
guide (according to these two naive strategies) for each window.