Avoid unnecessarily hashing key in _construct_guide_memoized()
This adds the argument `use_last` to
guide_search.GuideSearcher._construct_guide_memoized(). The method
typically hashes a key that can be large and slow to hash. When the
argument is set, it tells the method to use a previously found
inner_dict because the key is unchanged from the last call.
Effectively, this memoizes the inner dict of _memoized_guides.