Compress mostly contiguous indices in _memoized_guides
This adds a module, index_compress, to compress a collection of mostly
contiguous indices. It applies this to content stored in
guide_search.GuideSearcher._memoized_guides. There were two sets
in here that can be compressed well: seqs_to_consider (in the keys
of _memoized_guides) and covered_seqs (in the memoized output of
construct_guide()).

Additionally, this changes _construct_memoized_guide() to return a
set for covered_seqs (previously, a list) because this is needed
by the method that calls _construct_memoized_guide(). This required
many small changes to the guide_search test module (`[` to `{` and
`]` to `}`).