Only memoize guides constructed across a sufficiently high fraction of sequences
Many calls, particularly on highly diverse alignments, to
_construct_guide_memoized() may use a seqs_to_consider that represents a
small fraction of all sequences. In these cases, the same
seqs_to_consider is unlikely to be encountered again. Also, it should be
relatively fast to call construct_guide() on these seqs_to_consider.
These are reasons to avoid memoizing the result of construct_guide()
when seqs_to_consider represents a small fraction of all sequences in
the alignment.

This changes _construct_memoized_guide() to only memoize results when
seqs_to_consider is sufficiently large. It does this according to a
parameter, memoize_threshold, whose value is currently hard-coded.

The value of memoize_threshold can likely be tuned (higher values
lower memory and increase runtime, and vice-versa).