Do not memoize guides when designing primers
primer_search.PrimerSearcher uses guide_search.GuideSearcher to look for
primers in a way that makes it pointless to memoize the results of
construct_guide(). The actual memoization can take some time (and use
additional memory), so it would be best to not perform it. This commit
allows primer_search to tell guide_search to not memoize the results of
construct_guide().