Add option to design guides across all sequences in a target
By default, guides in an amplicon are designed only across
GUIDE_COVER_FRAC of the sequences that are bound by the designed primers
on each end. This makes sense as the guides should only cover the
sequences that would be amplified by the primers. However, this causes
the set of sequences that are designed against to change across windows
(i.e., amplicons or targets) as they may have different primers. As a
result, guide choices are less able to be memoized in
guide_search.GuideSearcher._memoized_guides across the windows.

This change adds an argument --guides-should-cover-all-seqs to
design.py, which causes guides to be designed for GUIDE_COVER_FRAC of
all the sequences in the alignment, rather than only those bound by the
primers. This means seqs_to_consider is more consistent across windows
and guide choices are better able to be memoized, which should improve
runtime on especially diverse input alignments.