Consider predicted activity when constructing guide, if set
This involves a large restructuring of construct_guide(). Previously,
the function considered (generally) just a guide from the largest
cluster. However, that guide might be predicted to be inactive against
all or most targets. Then, the function would not return a guide (or
would return a suboptimal one).

This rewrite has the function consider guides from all clusters, and
select the one that 'binds' (and is active against) the largest number
of sequences that need to be covered. The downside is that it is likely
to be slower.