Add data structure that shards k-mers across tries
This supports efficient specificity queries even with full sensitivity
to G-U base pairing. The general approach is described in the header of
the kmer_shard module. There are two implementations (one for p=1 and
another for p=2, as defined in that header):
kmer_shard.TrieSpaceOfKmersFullSig and
kmer_shard.TrieSpaceOfKmersSplitSig.

This commit also includes unit tests for the data structure.