Use serialized models directly for activity prediction
This commit considerably rewrites the predict_activity module.
Previously it used a crude approach to read models from another
directory. Now, this allows the module to read models saved in
TensorFlow's SavedModel format. It accepts paths to these models and
uses them for prediction.

This commit also changes the module to perform both classification and
regression -- i.e., it uses two models, one for each. For a guide-target
pair to be evaluated as 'True' (i.e., evaluate() returns True) it must
be both classified to be active and then determined to be highly active
based on regression.