Break GuideSearcher in base class and GuideSearcherMinimizeGuides subclass
This commit also generalizes some code in GuideSearcher, especially
having to do with memoization, to be more amenable to different kinds of
uses.

This will make it easier to have two versions of GuideSearcher for
different objective functions: one that minimizes the number of guides
and one that maximizes expected activity.