Replace serialized models with ones serialized in TensorFlow 2.2.0
The models were previously serialized with TensorFlow 2.1.0. It seems
there was an issue with those serializations (possibly a TensorFlow bug
stemming from changes to SavedModel made in 2.2.0): they do not load in
TensorFlow 2.2.0 or 2.3.0 using tf.keras.models.load_model(), even
though they should load according to TensorFlow's documentation on
SavedModel compatibility (https://www.tensorflow.org/guide/versions).

Below is a table listing, for each version used to serialize the models
(columns), whether they can be loaded in a version (rows):
         Serialized with:
         2.1   2.2   2.3
Loads
with:
  2.1    Yes   Yes   No
  2.2    No    Yes   No
  2.3    No    Yes   Yes

The models serialized with TensorFlow 2.2.0 and 2.3.0 do satisfy the
compatibility guarantee, but the ones serialized with 2.1.0 do not seem
to.

This commit replaces the models with versions now serialized using
TensorFlow 2.2.0, which should work in 2.2.0 and 2.3.0 (and hopefully
other minor versions within the same major release). There are no other
changes to the models.