Design with single reference sequence when no sequences pass curation
For 3 of ~1,900 vertebrate-infecting viral species, no sequences pass
curation (NCBI taxonomic IDs: 129956, 2015376, 50292). These each have
just 1-3 genome neighbors.

As one example, for taxonomic ID 129956 (Murine mastadenovirus A), there
is 1 genome neighbor (M22245) and the RefSeq is NC_000942. The RefSeq
is derived from M22245; however, for some reason, NC_000942 is ~31k
bp long while M22245 is just 4,333 bp long. As a result, M22245 does
not pass curation because it has too low overall identity to the RefSeq.

This commit tries to resolve these edges cases, in which no sequences
pass curation, by using a single RefSeq genome as input for design when
needed. It raises a critical message to indicate when this is the case.