Fix bug when writing target stats without predict activity model
When running in complete-targets mode without an activity model set
(e.g., for the minimize-guides objective), target_search.py would try to
compute expected activities for each guide. Doing so would raise an
error. This commit fixes that by using nan for the expected activity of
each guide in this case, as is the case for other activity prediction
stats.