Added diversity metric to design_naively; changed design_naively to get best n guides within a window
Allows design_naively to output designs based on a diversity metric (currently, only works with entropy). Added command line arguments to pick which methods are used, to design based on a reference sequence, and to set the number of guides to design for in each window. Modified constuct_guide_naively and find_guide_in_each_window in design_naively.py for the above and wrote a function to calculate entropy at each position (position_entropy) in alignment.py