Filter neighbors
Generalize add_metadata_to_neighbors_and_filter  to filter based on any of the metadata fields and add in taxonomic ID as a metadata field. This taxonomic ID indicates the most specific taxonomic ID given by the creators of the accession. It can be used to distinguish subspecies if subspecies is labeled on the accession (as in SARS-CoV-2).