Add option to write mean activity of guides in analysis
This adds an argument to analyze_coverage.py,
--write-mean-activity-of-guides, which writes to a file the mean
activity of a guide set across the targets. The commit also adds the
functionality in coverage_analysis.py to compute these values.

Note that the mean activity *excludes* primers -- i.e., it only
considers guides and does not account for whether the primers in a
design would amplify the target sequence. This is accounted for,
however, by the output of --write-frac-bound.