Add cas13a model argument; modify README
Add '--predict-cas13a-activity-model' as an argument so that users use the model even in packaged form. Rename folders containing Cas13a models to Cas13a for readability and generalizability; move to adapt/ directory so that it can be included as package_data. Modify README to account for these changes and explain you may need to download the FASTA file. Fix version.py to not reference CATCH and setup.py to include the models folder in the package. Code clean up.