Include consensus of all sequences as a representative guide
In addition to including the consensuses of clusters, we can consider
the consensus of all the sequences as a "representative" sequence. This
allows it to be placed in the ground set of candidate guides; it may
exhibit better detection activity, in some cases, than the consensus
of any of the individual clusters.