Consider consensus of all sequences when constructing optimal guide
In addition to considering the consensuses of clusters as an optimal
guide, we can consider the consensus of all the sequences. It may
exhibit better detection activity, in some cases, than the consensus of
any of the individual clusters.

This commit is similar to ea06477, but the function is used for a
different objective (minimizing guides rather than maximizing activity).