Tolerate short sequences in MinHash family
If an input sequence was too short (fewer than the number of k-mers
needed in a signature), the MinHash family previously terminated with a
message indicating so. This commit changes the behavior to allow such
sequences--repeating k-mer hashes in the signature as needed--while
giving a warning message that doing so may lead to unreliable distance
estimates.

This commit copies commit fc6e8a6 from the CATCH repository.