Log critical message to prevent an issue when using minimize-guides objective
With the minimize-guides objective, to detect a sequence a guide must be
within the specified number of mismatches (-gm/--guide-mismatches) even
when a predictive model is provided (when such a model is provided, both
the limit on the number of mismatches and the activity threshold must be
met). See search.OligoSearcherMinimizeNumber.construct_oligo() for where
these requirements are checked.

Effectively, using a mismatch limit helps to reduce the number of calls
to the predictive model and thereby improve runtime. However, it is not
needed -- down the road, the constraint on the number of mismatches
could be removed as long as a predictive model is provided.

The default value for -gm/--guide-mismatches is 0. Thus, if the argument
is not provided, minimize-guides will require 0 mismatches between
guides and targets. That is generally much too restrictive. The purpose
of the critical message added in this commit is to notify the user, when
-gm/--guide-mismatches is unset but a predictive model is, that the user
should consider also setting -gm/--guide-mismatches. To ignore the limit
on the number of mismatches in practice, it could be set to a
sufficiently high value (e.g., for a 28 nt guide,
`--guide-mismatches 10`).

The README already notes, for the argument -gm, that "this argument
serves as an additional requirement for evaluating detection on top of
the model".