Change default objective function to maximize-activity
Previously, the default objective function has been `--obj
minimize-guides`. This made things easier because it does not require a
predictive model. However, it is not what we generally use in practice
nor what we use in our publication.

This commit changes the default to be `--obj maximize-activity`. If a
predictive model is not set, design.py exits with a message to either
set one or to use `--use-simple-binary-activity-prediction`, which
resorts to using mismatches for binary predictions of activity.

This commit also updates the README accordingly.